requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
SQLAlchemy==2.0.28
asyncpg==0.29.0
requests==2.31.0
pytz==2024.1